+++ ompi/mca/mpool/base/mpool_base_mem_cb.c
@@ -47,7 +47,7 @@ void mca_mpool_base_mem_cb(void* base, size_t size, void* cbdata,
     }
           
     base_addr = down_align_addr( base, mca_mpool_base_page_size_log);
-    bound_addr = up_align_addr((void*) ((unsigned long) base + size - 1), mca_mpool_base_page_size_log);
+    bound_addr = up_align_addr((void*) ((ptrdiff_t) base + size - 1), mca_mpool_base_page_size_log);
     for(item = opal_list_get_first(&mca_mpool_base_modules);
         item != opal_list_get_end(&mca_mpool_base_modules);
         item = opal_list_get_next(item)) {
