+++ ompi/mpi/c/type_set_name.c
@@ -48,7 +48,7 @@ int MPI_Type_set_name (MPI_Datatype type, char *type_name)
   }
 
   memset(type->name, 0, MPI_MAX_OBJECT_NAME);
-  length = strlen( type_name );
+  length = (int)strlen( type_name );
   if( length >= MPI_MAX_OBJECT_NAME ) {
     length = MPI_MAX_OBJECT_NAME - 1;
   }
